finished backend

+++ dao/transaction.go
@@ -7,6 +7,6 @@ type Transaction struct {
 	UserId    int       `json:"user_id"`
 	StockId   int       `json:"stock_id"`
 	SellAmt   int       `json:"sell_amt"`
-	SellPrice int       `json:"sell_price"`
+	SellPrice float32   `json:"sell_price"`
 	PostDate  time.Time `json:"post_date"`
 }
